requirements.txt
python-telegram-bot
pika
SpeechRecognition
Pillow
pytesseract
python-dotenv
pydub